Add project deploy script

+++ publish.ps1
@@ -1,2 +1,5 @@
-Get-ChildItem publish | Remove-Item -Recurse
+# publish the project
+if ((Test-Path .\publish) -eq $True) {
+    Get-ChildItem publish | Remove-Item -Recurse
+}
 dotnet publish -o publish
